requirements.txt
pandas~=1.1.4
joblib~=0.17.0
xgboost~=1.2.1
numpy~=1.19.4
kfp~=0.5.1
google-cloud-storage~=1.29.0